csrf

+++ database/deleteBlog.php
@@ -1,7 +1,14 @@
 <?php 
-    if ($_GET) {
-        require_once "DB.php";
-        $db = new DB();
-        $db->deleteBlog($_GET["id"]);
+    session_start();
+    if ($_SESSION) {
+      if ($_SESSION["loggedIn"]) {
+            if ($_GET) {
+                require_once "DB.php";
+                $db = new DB();
+                $db->deleteBlog($_GET["id"]);
+            }
+        }
+    } else {
+        header("location: ../admin/login.php");
     }
 ?>
